% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_source_win.R
\name{build_source_win}
\alias{build_source_win}
\title{Build the window to read-in source data}
\usage{
build_source_win()
}
\description{
This function is run when a user clicks the "Load source data" button in the
main MixSIAR GUI, and creates a separate GUI window (\code{source_win}) where
the user loads the source data. \code{build_source_win} calls
\code{\link{load_source_data}} when the user clicks the "I'm finished" button.
}
\details{
The "Load source data" window contains:
\enumerate{
  \item Questions the user if their source data is "by factor" (if they have Fixed or Random Effects),
  \item Questions the user if they have Concentration Dependence data,
  \item Buttons to either load "raw" or "means + SD" source data.
}

If \code{mix} does not yet exist, a WARNING appears.

A green check appears if the source data is successfully loaded, or a red_x
if not.
}
\seealso{
\code{\link{load_source_data}}, which is run when the "Load source data"
         window is closed by clicking the "I'm finished" button.
}
