% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mixsiar_envir.R
\docType{data}
\name{mixsiar}
\alias{mixsiar}
\title{mixsiar}
\format{An object of class \code{environment} of length 0.}
\usage{
mixsiar
}
\description{
\code{mixsiar} is an R environment the GUI version of MixSIAR uses to store
objects and pass values through functions. Internal---only used
in the event a user wants to extract/see MixSIAR objects (e.g.
loaded mix data are accessed via \code{mixsiar$mix}, and the finished
model output, \code{rjags} object, is \code{mixsiar$jags.1})
}
\keyword{datasets}

