% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot prior}
\usage{
plot_prior(alpha.prior = 1, source, plot_save_pdf = TRUE,
  plot_save_png = FALSE, filename = "prior_plot")
}
\arguments{
\item{alpha.prior}{vector of alpha (dirichlet hyperparameters, none can be = 0)}

\item{source}{output from \code{\link{load_source_data}}}

\item{plot_save_pdf}{T/F, save the plot as a pdf?}

\item{plot_save_png}{T/F, save the plot as a png?}

\item{filename}{name of the file to save (e.g. "prior_plot")}
}
\description{
\code{plot_prior} plots your prior on the global diet proportions (p.global)
and the uninformative prior side-by-side. Your prior is in red, and the
"uninformative"/generalist prior (alpha = 1) in dark grey.
}

